OC-595 ~ Removed DistributionAutomation MeterReading logging
The MeterReadings are a left-over from a hackathon that has taken place.

The Avro format mentioned in the story acceptance criteria is one that
still has to be defined as an outcome of OC-574. Once that is done, the
peak shaving consumer tests that work with a String based consumer
record value, will have to be updated to use the new Avro format.