OC-578 ~ Updates some code towards MQTT peak shaving Cucumber tests
Adds some not yet implemented steps to the peak shaving feature
scenario.

Moves version number for Cucumber dependency to the MQTT simulator to
the integration tests parent POM.

Adds the device identification in the step to send a measurement report
so it can be used as part of the topic name.

Adds PAYLOAD to PlatformDistributionAutomationKeys.

Adds mqtt.simulator.startClient to the MQTT simulator properties.

Removes some empty lines from the simulator spec JSON file.

Refactors Broker to get rid of a Sonar issue modifying the static value
of running from an instance method.

Renames LnaClient to SimulatorSpecPublishingClient.