OC-607 ~ Makes Microgrids XSDs accessible by URL
Retrieving the adhocmanagement XSD from the MicrogridsAdHocManagement
WSDL by the expected URL fails with a 405 HTTP response.

This is a result of the SimpleXsdSchema bean name that did not match the
name of the XSD file (bean name is "adhocmanagement-ws-microgrids").
To avoid issues with duplicate bean name errors, the XSD has been
renamed to match the bean name. The same goes for the "common.xsd"
schema.

The URL for accessing the XSD should now be what is to be expected from
looking at the XSD import in the WSDL.