FLEX-5300 ~ Adds 60870-5-104 LMD connection monitoring
Makes sure connections are kept active for the light measurement devices
with protocol IEC 60870-5-104.

Adds DeviceConnection104LmdScheduledTask, a scheduled task that ensures
connections with (the gateways of) the IEC 60870-5-104 light measurement
devices.

Changes may probably be required when requests to the IEC 60870-5-104
light measurement devices will be handled from the web service layer.
This has not been addressed for FLEX-5300 since another item already is
on the backlog to make sure the correlation UID is kept between request
and response (FLEX-5349), which needs to be dealt with before the web
service responses can be matched with requests at all.