FLEX-5453 ~ Adds test annotations and splits method in two
Some of the unit test methods supposed to test functionality missed the
Test annotation, which has been added to have them executed.

The method getMostRecentDeletingOutdatedPendingFirmwareUpdates has a
long name, revealing that it does multiple things. This is refactored
into two more specialized methods
(getMostRecentSsldPendingFirmwareUpdate and
deleteOutdatedSsldPendingFirmareUpdates).

The id of the SsldPendingFirmareUpdate is a Long object value. Therefore
the use of "!=" has been replaced by "!Objects.equals".