FLEX-5453 Prevent error getting firmware version with multiple pending firmware updates (#419)
* FLEX-5453 ~ Deals with multiple SSLD pending firmware updates

Returns a list of SsldPendingFirmwareUpdates searching by device
identification and deals with multiple values being returned.

* FLEX-5453 ~ Adds unit test for handling the lists of pending updates

Adds unit tests for the code that worked with a single
SsldPendingFirmwareUpdate for a given device identification, that has
been changed to deal with lists of pending updates instead.

* FLEX-5453 ~ Adds test annotations and splits method in two

Some of the unit test methods supposed to test functionality missed the
Test annotation, which has been added to have them executed.

The method getMostRecentDeletingOutdatedPendingFirmwareUpdates has a
long name, revealing that it does multiple things. This is refactored
into two more specialized methods
(getMostRecentSsldPendingFirmwareUpdate and
deleteOutdatedSsldPendingFirmareUpdates).

The id of the SsldPendingFirmareUpdate is a Long object value. Therefore
the use of "!=" has been replaced by "!Objects.equals".

* FLEX-5453 ~ Fixes spelling mistake and splits act and assert statements

* FLEX-5453 ~ Splits act-assert line in two lines in two methods

Co-authored-by: Sander van der Heijden <sander.van.der.heijden@cgi.com>