FLEX-5492 ~ Adds missing message processor and uses constants.
- The message processor class added in this commit prevents a 'no
message processor found for message type' exception. There is some
default error handling in that case, which sends the message along to
the domain component. To prevent exceptions in the logging and depending
on some kind of default error handling behavior, is it better to
explicitly handle the message type with a message process class.
- Instead of using some hard coded string, use the enumeration members
of the MessageType enumeration. This will make refactoring of the code
in the future less error prone.