Fixed several SonarQube issues
One main issue was with errors being thrown within a catch, and then being catched again and send to a method. I changed it so the first catch calls the method with the new exception. This is also explained in a comment in DeviceManagerEndpoint. Due to the amount of occurences I felt that I needed to explain what I did and why in case there is a good reason for doing it the way it was done