SLIM-2675 ~ Fixes core device installation Cucumber tests
Removes earlier attempt to add an organization as a given in the
StartDevice and StopDevice incorrect parameter scenarios.
The main reason being that it does not seem to make sense to add the
organization when the examples include an identification like
ORGANIZATION_ID_UNKNOWN.
With the other changes, it looks like the tests would also succeed when
the organization is added though.

Only update the device with a new DeviceAuthorization if an organization
exists (not only an organization identification, which may be for an
unknown organization). It does not make sense to create a device
authorization without an organization.

Change the way the domains for a new organization are checked for an
empty value to avoid null pointer exceptions.