FLEX-5527 ~ Moves 'handle schedule' call in OSLP channel handler server to a more robust location.
Tests have shown that the introduces call to set schedule is too soon. This causes problems with the device simulator, optimistic locking exceptions caused by the fact the register device and confirm register device steps are not completed yet. This change in the execution order seem to deliver a more robust solution.