FLEX-5527 ~ Makes sure exceptions, like connection refused or connection time-out, which result in an IOException, are propagated.
Testing revealed a major bug, which causes exceptions to be 'ignored'. Those exceptions were not handeled and converted to a response sent by protocol adapter to the platform components. Calling the response handler when an exception occurs and introducing a try-catch when bytes are written to the channel solves these problems.