FLEX-5537 ~ Enables OslpDecoder to decode multiple OslpEnvelopes
Adds a test with a channel handling input for multiple OslpEnvelope
values.
Updates the OslpDecoder not to break after processing the first
OslpEnvelope. Consecutive messages can be decoded now the DONE state is
removed. Output of the decoded message and a reset of the decoding
related fields is implemented as part of the last step that actually
processes some input.

Changes the OslpDecoder to be a little more in line with the example of
the ReplayingDecoder with enum state from the javadocs.
Avoids reading or updating of the reader index of the ByteBuf that is
decoded, leaving these details with the base class provided by the
library.