FLEX-5534 ~ Fixes error in find events.
When all events for all devices are queried, the device identification is null. This causes an error when searching for device authorizations. Fixed the subquery by changing the type from Long to String and made sure the device authorizations are queried using the device foreign key, instead of the device identification.