FLEX-5537 ~ Makes hex dump logging in Netty LoggingHandler configurable
Removes the OslpLoggingHandler, as logging the hex dump is configurable
in a recent Netty LoggingHandler by specifying a ByteBufFormat. Being
able to turn hex dump logging on or off was the only reason that the
OslpLoggingHandler was created.

Instead of needing to change the code to turn hex dump logging on or off
some settings are introduced to configure the logging level and whether
to include hex dumps.