FLEX-5580 ~ Removes duplication in measurement to Analog conversion
Measurement values are converted to an Analog list using an Enum that
has a definition describing each position in the list of values. This
definition combined with the matching value in the measurement values
list can produce an Analog value.

The earlier implementation had a StringArrayToAnalogList converter for
each of the Enums describing the values. These changes refactor those
converters into a single class that can use any of the Enums that
matches the number of values to be converted.