Fix comments on issues in ClockStatus and ClockStatusBit.
There were 2 issues for which changes were requested. The first one is suppressing the issue of returning null in ClockStatus. The method where this issue is suppressed, is never called and therefor it was changed to return an empty set.

The other issue was returning null in ClockStatusBit. Here we looked at ways to fix it, but kept running into the problem that the code was written with the idea that certain values could / should be null. In the end we decided to neither change nor suppress the issue, since someone with more knowledge about the specific code should look at it instead.