FLEX-5537 Investigate OSLP messages with unknown device UID (#477)
* FLEX-5537 ~ Adds OslpDecoder unit tests

Adds unit tests to verify that the OslpDecoder works properly, even if
the necessary bytes come in unexpected patterns (like partial chunks,
simulating some kind of bad connection).

* FLEX-5537 ~ Enables OslpDecoder to decode multiple OslpEnvelopes

Adds a test with a channel handling input for multiple OslpEnvelope
values.
Updates the OslpDecoder not to break after processing the first
OslpEnvelope. Consecutive messages can be decoded now the DONE state is
removed. Output of the decoded message and a reset of the decoding
related fields is implemented as part of the last step that actually
processes some input.

Changes the OslpDecoder to be a little more in line with the example of
the ReplayingDecoder with enum state from the javadocs.
Avoids reading or updating of the reader index of the ByteBuf that is
decoded, leaving these details with the base class provided by the
library.

* FLEX-5537 ~ Reduces number of lines in OslpDecoder switch case

Reduces the number of lines in the OslpDecoder switch case for the
PAYLOAD_MESSAGE based on a SonarQube warning.
Refactors release of temporary ByteBuf in readBytes, so it always
happens, even with unexpected exceptions from getBytes.

* FLEX-5537 ~ Makes hex dump logging in Netty LoggingHandler configurable

Removes the OslpLoggingHandler, as logging the hex dump is configurable
in a recent Netty LoggingHandler by specifying a ByteBufFormat. Being
able to turn hex dump logging on or off was the only reason that the
OslpLoggingHandler was created.

Instead of needing to change the code to turn hex dump logging on or off
some settings are introduced to configure the logging level and whether
to include hex dumps.

* FLEX-5537 ~ Changes OslpChannelHandlerServer to fix memory leak

OslpChannelHandlerServer no longer tracks channels to send signed
responses in a map that retains all its entries that were ever added.

Adds a ChannelCache where cached channels are removed when they are
requested. The cache uses an expiration time-out in milliseconds, after
which cached entries become eligible to be removed.
The cache allows a channel to be cached multiple times and retrieved
that same amount of times before it is removed.

Closes the ChannelHandlerContext in scenario's where a proper response
will never be sent from the OslpChannelHandlerServer.

Removes ctx.fireChannelRead(message) from the OslpChannelHandlerServer
as there is no later handler in the pipeline to do anything with the
OslpEnvelope.

* FLEX-5537 ~ Adds unit tests for the ChannelCache

* FLEX-5537 ~ Makes ChannelCache expiration time configurable