Fixes timing issues in ChannelCache tests
With FLEX-5537 a ChannelCache was introduced as a replacement of a map
of channels.

The unit tests simulated a delay for the expiration time of the cache in
a way that was sensitive to timing issues when the tests were executed
using multiple threads under heavy load.

This changes the ChannelCache to depend on a Clock (by default just the
system clock) that can be controlled during the unit tests. This way
delays (passing of time) can be simulated in the tests without actually
resorting to sleeping or delayed execution with an executor.