FLEX-5474 ~ Adds reboot behavior to the OSLP web device simulator
When the OSLP web device simulator simulates a reboot a configurable
delay in seconds is applied before the device registration messages are
sent. During this delay the Netty channel for new requests will be
disconnected by the OlspChannelHandler.

The delay has a default value from configuration (reboot.delay.seconds)
and can be changed from the devices page in the web-device-simulator.

Fixes a number of warnings in the DeviceManagementController.
The general RequestMapping with a method is replaces by GetMapping or
PostMapping.
Some checks are done in the API methods that return a response for a
bad request in case there is something wrong with the provided data.
Some of the methods that were altering state have been renamed from get
to set.
The random for generating bytes is placed in a field instead of creating
a new one on each call in the method where it is used.
An incorrect logger statement with an exception is fixed.

The labels in the device list page have been placed in actual label
elements for the controls they belong with (making clicking the check
box inputs simpler as one side effect).

A call to fireChannelRead is removed from the OslpChannelHandler as it
is the last handler in the pipeline. Passing the message on beyond the
last handler triggers some debug level warning messages from Netty.