LSM-6 ~ Adds low voltage message signing and verification
Adds MessageSigner to be configured with at least one of two keys: a
private key for message signing or a public for signature verification.
This MessageSigner is used with the producer and the consumer of Kafka
Avro messages for low voltage.

The producer only needs the private signing key, but in the test setup
the public verification key is configured to be able to include the
verification in the test scenario.
The consumer only needs the public verification key, but in the test
setup the private signing key is configured to be able to include the
verification in the test scenario.

Updates the Kafka low voltage message producer to sign the Avro messages
that it sends to its topic.
Updates the Kafka low voltage message consumer to log warnings if it
notices a consumed message can not be verified.