FLEX-5349 ~ Processes review comments
Adds some text in an attempt to provide an explanation why all devices
behind a gateway need to be iterated over sending a light sensor status
response.

Small change to remove correlationUid faster when an exception occurs,
as well as visually closer to enqueuing, taking this responsibility out
of the method for sending the not supported exception.

Removes the captor for the correlation UID String in the tests, and uses
the known correlation UID instead.