FLEX-5349 ~ Uses IP address of the gateway on 104 LMD get status
The get status request for the public lighting AdHocManagementService
gets sent with the OSGP core request message sender using the IP address
of the gateway (if the device has a gateway device) instead of the IP
address of the device itself (which is null for a device behind a
gateway).

It would probably be a good idea to change this in more or even all
places, depending on how generic gateway functionality should work.
For now this fix is only done to make get status work for 104 LMDs in
the setup with a gateway (without the need to duplicate the gateway's IP
address with all LMDs behind it).