OC-718 ~ Adds helper class to execute ICMP echo requests
Adds Pinger to the shared networking package, as a helper to perform
ICMP echo requests.

The Pinger uses a CommandBuilder to create the commands to be executed,
customizing command flags to match the ping flavor expected for the
operating system the code is executed on.

The Pinger uses a CommandExecutor to execute the ping command, where a
timeout may be provided to make sure a value (false) is returned when
ping results are not processed before the timeout expires.