OC-718 ~ Adds configurable device pinging to DLMS protocol adapter
When a DLMS device connection is about to be created the code in the
DlmsConnectionHelper checks whether pinging is enabled by configuration.
If pinging is enabled, and an IP address is known for the device to be
connected, a Pinger according to configuration is used to send a ping
echo message to the device.

Removed the assertion on the result of toString of the ThrottlingService
from DeviceRequestMessageListenerIT. The string did not match, but it
was not clear why this should be asserted on.