FDP-114 ~ Synchronizes SMR 5 invocation counter without retries
Avoids resetting the invocation counter and throwing an exception to
deal with the invocation counter initialization utilizing the GXF retry
scheduling mechanism.

This should work assuming that a connection to the management client of
the DLMS device can be set up immediately after gracefully disconnecting
from the public client in case the public client connection is utilized
to retrieve an up to date value of the invocation counter.

Resetting the invocation counter (by setting it to null in the database)
is no longer needed as the connection will be retried with an explicit
flag to initialize the invocation counter that bypasses the check in the
database on the second attempt.