FDP-114 ~ Adds a delay between DLMS connections to the same device
Adds a configurable delay in seconds that will be applied between
setting up consecutive DLMS connections to the same device.
This should prevent SocketException with message "Connection reset" that
appears to occur when a TCP connection is built for the same target host
and port of a connection that was just closed.

Removed the code to gracefully disconnect the public client connection,
falling back to just closing it again. The disconnect did not appear to
fix the problem with setting up the next connection on the DLMS
management client of the device. The issue with the repeated connection
looks like to be caused client side setting up the TCP connection, and
should be mitigated by an added delay before attempting to set up the
connection.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>