FDP-114 ~ Increases reconnect delay updating DLMS invocation counter
Increased the DLMS delay between consecutive from 10 to 20 seconds in
the properties file, because at 10 seconds the flow was observed to fail
in some test case due to the "Connection reset" SocketException when
building the new TCP connection.

Removed the DlmsDeviceRepository from the InvocationCounterManager and
tests. Saving the device with an updated invocation counter leads to an
exception later-on in the DlmsConnectionMessageProcessor where the
method doConnectionPostProcessing saves the device with the updated
invocation counter adding the count for the device communication that
was just performed.
The exception was something like "OptimisticLockException: Row was
updated or deleted by another transaction".
There is no need to call explicit saves on the device in the
InvocationCounterManager as it will be saved with the final update on
the invocation counter from a transaction in the
DlmsConnectionMessageProcessor anyway.

Removes resetInvocationCounter from InvocationCounterManager as it is no
longer used, and is not needed with the updated invocation counter
handling.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>