FDP-111 ~ Updates code dealing with metadata and max schedule time
Adds SoapHeaderMessageMetadataInterceptor to gather information from the
SOAP headers in a single MessageMetadata object instead of in several
different elements each with their own parameter.
Configures the interceptor in the Smart Metering WebServiceConfig, so it
can (initially) be used by the smart metering bundle endpoint.

Changes Flyway script to add max schedule time:
 * checking the column does not exist before adding it;
 * adding a comment on the new column;
 * changing the type to match that of the existing time based column.
 * renaming max_scheduled_time to max_schedule_time to be more in line
   with the name used elsewhere and to indicate that this is something
   about future scheduling and not something already scheduled.

The ScheduledTaskExecutorService now handles the IP address taking a
possible gateway device into account matching what is done in the
DeviceResponseMessageService, allowing scheduled tasks for a device
behind a gateway to target the IP address of that gateway.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>