FDP-111 ~ Sets complete MessageMetadata building ProtocolResponseMessage
Calling both messageMetadata and methods like domain on the builder of
the ProtocolResponseMessage could cause values to be overwritten when
called in the wrong order. To prevent this, and to create a stronger
separation between the message metadata and other data, the message
metadata is wrapped and set in a single call as much as possible.

Reverts some changes to the RequestMessage, ProtocolRequestMessage,
ResponseMessage and the ProtocolResponseMessage. Do not remove any
fields, avoiding incompatible serialization changes and for now keep the
same serialVersionUID to prevent problems with serialized messages that
could still be in the queues or the response data tables.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>