FDP-111 ~ Send smart metering messages using object and metadata
Sends smart metering requests from domain to core using the serializable
object and message metadata as arguments. No longer uses RequestMessage
as intermediate structure, as it does not have all the metadata
elements.

Creates request and response messages for smart metering with a single
call to configure the message metadata as much as possible.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>