FDP-111 ~ Fixes Cucumber tests after breaking bundle step
The step "the bundle request is received" has been extended to accept
SOAP header element data from the step table. This broke the existing
scenarios where no header data was specified.
This reverts the original code behind the step and adds a new step
"the bundle request is received with headers" for bundle requests to be
sent with SOAP header elements.

Changes the definition of ScheduleTime and MaxScheduleTime in
base-ws-smartmetering.xsd from xsd:dateTime to xsd:long as the code
parses these values as Long.
Aligns the references in the XSD so that ScheduleTime and
MaxScheduleTime are treated the same.
Although these elements do not appear to affect validation, they are
left in for documentation purposes.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>