FDP-111 ~ Invalidates any blank IP address instead of only null
The code in the DlmsConnector checks the IpAddress set with the device
for communication and throws a FunctionalException if it is known to be
invalid because it is empty.
The check for emptiness looked for the value null, but not for empty or
blank text. Code with the MessageMetadata returns an empty String for
the IP address if it is not set.
Now validation throws the exception for any blank value for the IP
address.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>