FDP-109 ~ Fixes issue with discarding permits
The implementation for discarding a permit removed the permit to be
discarded from the database. With these changes the in memory permit
count for the network segment the permit had earlier been granted for is
decreased as well.

To be able to run the throttling service application the database driver
should not have a test scoped dependency.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>