FDP-109 ~ Returns definite in memory count of max concurrency
MaxConcurrencyByThrottlingConfig has an updateMaxConcurrencyFromDatabase
method that could not actually update the in memory max concurrency
value if a concurrent registration of the throttling configuration has
updated the map.

Fixes formatting in PermitsByThrottlingConfig that caused the build to
fail during the check action of the fmt-maven-plugin.
In order to always return the last updated in memory value, the
maxConcurrency as read from the database is not returned directly, but
the current updated value actually stored in maxConcurrencyByConfigId is
returned instead.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>