FDP-109 ~ Passes BTS ID and Cell ID via message metadata
Similar to how the IP address was transferred from higher layers to the
protocol adapter, the base transceiver station ID and cell ID are passed
via message properties as well, and used with the calls to request a
permit to access the network.

Refactored the ThrottlingPermitDeniedException to also be useable when
the network segment IDs (base transceiver station ID and cell ID) are
not set.
Moved the ThrottlingPermitDeniedException to the throttling client
project to simplify the call flows from the protocol adapter (no need to
make different calls when the network segment IDs are available or not).

Sets throttling.client.enabled=false by default in
osgp-adapter-protocol-dlms.properties, because the throttling is not
available (yet) in the environment the build server works with.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>