FDP-109 ~ Fixes bug with bts_id and cell_id -1 for throttling
The ProtocolRequestMessage did not properly keep the values for base
transceiver station ID and cell ID from the MessageMetadata, causing the
default value of -1 to be used for the network segment with throttling,
regardless of the actual values in the database.

Adds ConnectionThrottling.feature that is skipped on the build server
that has a couple of scenarios that may come in handy when investigating
the throttling service application used with the protocol adapter DLMS
on a development machine. For enabling such testing on a build server
the platform would have to be deployed with separate configuration,
enabling the throttling client.

Adds notifications back on the queue within the NotificationService for
the Cucumber tests, because otherwise notifications get lost in the
throttling scenario with requests for multiple devices.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>