FDP-109 ~ Updates permit clean up job to release from memory
The permit clean up job used to delete expired permits directly from
the database. Now it goes through the SegmentedNetworkThrottler in order
to make sure the permits are deleted from the in memory structure as
well as from the database.
Some investigation is still needed because an initial test starting the
application with some expired permits inserted in the database removed
the permits from the database, but from the flow where
segmentedNetworkThrottler.releasePermit returned false.

Removed the unused autowired ClientRepository from PermitCleanUpJob.

In order to be able to override application properties and logging
configuration from the environment, packaging was changed to war and a
context.xml file was added allowing to specify overrides through the
naming context. In order for this to work, among other changes, the
application.properties had to be renamed to
osgp-throttling-service.properties.

The earlier ThrottlingService that was replaced by the ThrottlingClient
when the client is configured to be enabled, used a timer that causes a
reset of a timer lock including logging. Because either the
ThrottlingClient or the ThrottlingService is active, the
ThrottlingService is made conditional on the
ThrottlingServiceEnabledCondition, which is the inverse of the
ThrottlingClientEnabledCondition.

Some properties that were configured with defaults in the code have been
added to osgp-throttling-service.properties with some description, to
make it easier to see what can be configured deploying the
ThrottlingServiceApplication, without having to go through the code.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>