FDP-109 ~ Changes expired permit clean up to use batching
Changes clean up to work with bacthes of configurable size to prevent
issues if there would be an unexpected large number of permits that
should be cleaned up.

Alters initialization of overridden logging configuration in the
SpringApplicationBuilder with the ThrottlingServiceApplication, in order
to make it actually work now the application is deployed as war.

Removes configuration of the spring-boot-maven-plugin from the POM,
because the build artifacts do not have to be repackaged for standalone
execution now the application is deployed as a war file.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>