FDP-109 ~ Improves robustness of smart meter test NotificationService
Tests, for instance from the SystemEvent feature waiting to verify no
more system events would be notified could fail by notifications from
other test cases (like a pushed alarm notification).

The check for lack of notifications of system events now only checks
system event notifications.

When waiting for notifications of a certain type, notifications of other
types are put back on the queue, like when waiting for a correlation UID
and receiving a notification for another correlation UID.
Putting a notification back on the queue is delayed a little, giving the
wait loop some time for the notification that actually matters in the
test assertion.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>