FDP-109 ~ Removes sort order selecting expired permits
Removes the sort order from the page request to obtain expired permits.
Sort order is not imported during clean up, only the first page is ever
requested, for which the size limiting the total number of results is
enough.

Since paging beyond the first is never needed, just returns a list
instead of a page of permits.

Adds an index on the created_at column of permit, to help with the query
selecting permits with created_at before a certain point in time.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>