FDP-109 ~ Improves test robustness notifications
An earlier improvement filtered the notifications by type to make sure a
system event was handled. Before filtering it turned out a push
notification alarm could make the test fail.

This updates the Cucumber test dealing with push notification alarms to
not only check if the pushed alarm ends up in the logging database, but
also check a notification was sent and the push notification alarm
response can be retrieved.
By including retrieval of the pushed alarm message, its notification
should no longer be able to break other test cases.

Processing of the pushed alarms still had some notification check issues
due to the way pushed alarms are treated in the
PushNotificationAlarmMessageProcessor from core. When two alarms by the
same devices arrive in the queues at about the same time, processing can
be done in multiple concurrent listener threads, trying to update the
same device data from the same original version. The manner in which the
resulting StaleObjectStateException is handled appears to lead to more
notifications for the same alarm, which in turn leads to test assertion
failures regarding these notifications.
The changes to the test code are not a real fix, but treat the symptoms
by adding a few seconds between simulated alarms from the same device,
which should prevent the concurrent processing of multiple alarms for a
single device.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>