FDP-370 ~ Moves logic to get an IP address before connecting to a device
The logic to obtain an IP address was called from a finder method, which
has led to confusion and unnecessary calls to a session provider for
devices that do not have a static IP address.

Obtaining an IP address is removed from the finder and placed in the
code closer to where the actual connection is set up (in the
DlmsConnectionFactory and RecoverKeyProcess).

Adds tests confirming the IP address is set from the new location in the
code for devices with static or non-static IP address.

Fixes some unit tests that were broken by unfinished work from an
earlier commit.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>