FDP-580 ~ Makes DA MQTT to Kafka flow more generic
The flow for domain distribution automation where data is received from
an MQTT broker (to which devices publish their measurement data) and
published to Kafka is made more generic.
The text message from MQTT is left in String form and published as such
to a Kafka topic.
This way very specific handling to a single customized use case is
removed from the code, leaving the GXF implementation more open.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>