FDP-580 ~ Adds clean session and keep alive to simulator MQTT client
Adds attributes to use clean session and keep alive settings with the
MQTT simulator client connections.
This should help in preventing error logs like:

  Client SimulatorSpecPublishingClient received Ack CONNACK
  Client SimulatorSpecPublishingClient started
  Exception while starting client.
  com.hivemq.client.mqtt.exceptions.MqttSessionExpiredException:
    Session expired as connection was closed.

when publish is called from the onConnect method in
SimulatorSpecPublishingClient handling messages to be published.

Catches BindException in the bean method simulator from SimulatorConfig,
which was seen on Spring application context reloads where a broker was
already running. This could use some better actual fix in the way this
is set up, which I did not easily see to be quickly implemented.
If the current hack appears to leave open issues, this will still need
to be properly addressed.

Adds some more specific and additional logging to the MQTT Client and
SimulatorSpecPublishingClient that should help a little to get a clear
picture of what is exactly happening should things go wrong.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>