FDP-580 ~ Fixes MQTT simulator broker config
Separates the MQTT simulator for handling MQTT client actions from the
MQTT broker that handles the server side.
The broker is configured as a separate bean that starts and stops
independent of the simulator.
This enables deploying the simulator project in a servlet container to
get an MQTT broker, to which different clients can be applied to handle
behavior as desired, for instance with Cucumber tests.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>