FDP-580 Make Distribution Automation MQTT to Kafka flow more generic (#784)
* FDP-580 ~ Makes DA MQTT to Kafka flow more generic

The flow for domain distribution automation where data is received from
an MQTT broker (to which devices publish their measurement data) and
published to Kafka is made more generic.
The text message from MQTT is left in String form and published as such
to a Kafka topic.
This way very specific handling to a single customized use case is
removed from the code, leaving the GXF implementation more open.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>

* FDP-580 ~ Updates POMs

Moves some configuration for test plugins to the general GXF platform
super POM.
Fixes the version of the dependency plugin so that a recent version will
be used with for instance dependency:analyze.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>

* FDP-580 ~ Adds clean session and keep alive to simulator MQTT client

Adds attributes to use clean session and keep alive settings with the
MQTT simulator client connections.
This should help in preventing error logs like:

  Client SimulatorSpecPublishingClient received Ack CONNACK
  Client SimulatorSpecPublishingClient started
  Exception while starting client.
  com.hivemq.client.mqtt.exceptions.MqttSessionExpiredException:
    Session expired as connection was closed.

when publish is called from the onConnect method in
SimulatorSpecPublishingClient handling messages to be published.

Catches BindException in the bean method simulator from SimulatorConfig,
which was seen on Spring application context reloads where a broker was
already running. This could use some better actual fix in the way this
is set up, which I did not easily see to be quickly implemented.
If the current hack appears to leave open issues, this will still need
to be properly addressed.

Adds some more specific and additional logging to the MQTT Client and
SimulatorSpecPublishingClient that should help a little to get a clear
picture of what is exactly happening should things go wrong.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>

* FDP-580 ~ Fixes MQTT simulator broker config

Separates the MQTT simulator for handling MQTT client actions from the
MQTT broker that handles the server side.
The broker is configured as a separate bean that starts and stops
independent of the simulator.
This enables deploying the simulator project in a servlet container to
get an MQTT broker, to which different clients can be applied to handle
behavior as desired, for instance with Cucumber tests.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>

* FDP-580 ~ Resolves review comments

Treats String message consumer and producer files as new and updates the
copyright year to 2022.

Removes throws clause from bean method for IOException that is no
longer required to be declared.

Refactors StringMessageProducerTest extracting some sections of the test
into separate methods.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>
Signed-off-by: Sander Verbruggen <sander.verbruggen@alliander.com>